+++ pkg/tracker/channel/init_test.go
@@ -10,4 +10,5 @@ func TestInit(t *testing.T) {
 	assert.NotZero(t, len(socialChannel))
 	assert.NotZero(t, len(shoppingChannel))
 	assert.NotZero(t, len(videoChannel))
+	assert.NotZero(t, len(aiChannel))
 }